tests: share the udmabuf helper between the convert test and fuzzer
extract the sealed-memfd -> udmabuf wrapping into tests/udmabuf_util.h so
test_convert.c and fuzz_convert.c stop redefining struct udmabuf_create and
repeating the memfd/ftruncate/seal/UDMABUF_CREATE sequence. also add a blank
line after each changelog subsection heading and record the untrusted-convert
SIGBUS gotcha in the checklist.
