+++ CHANGELOG.md
@@ -7,6 +7,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
 ## [0.4.12] - 2026-07-20
 
 ### Security
+
 - `drmtap_convert_dmabuf()` hardened at the untrusted convert boundary. A
   coverage-guided fuzzer found that a descriptor claiming a frame larger than
   the fd actually backs made the CPU fallback mmap and read past the fd,
@@ -19,6 +20,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
   -EINVAL instead of faulting.
 
 ### Added
+
 - `fuzz/fuzz_convert.c` (+ build.sh, README): a libFuzzer target for the
   convert boundary, driving hostile descriptors over real udmabuf-backed
   dma-bufs. It survives tens of millions of runs clean after the fix.
@@ -28,6 +30,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
 ## [0.4.11] - 2026-07-20
 
 ### Hardening
+
 - `libdrmtap.so` now exports only the public API. A linker version script
   (`libdrmtap.map`, anonymous node so the public symbols stay unversioned) keeps
   every internal cross-module symbol local: `drmtap_set_error`,
@@ -43,6 +46,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
   privileged process opens.
 
 ### Changed
+
 - Built as both a shared object (version-scripted, installed, the one loaded via
   dlopen) and a static archive (full symbol set). The white-box unit tests link
   the static archive; examples link the public shared library.
@@ -50,6 +54,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
 ## [0.4.10] - 2026-07-20
 
 ### Added
+
 - `drmtap_grab_desc()`: the exporter-side counterpart to
   `drmtap_convert_dmabuf()`. Does a zero-copy grab and fills a complete
   `drmtap_dmabuf_desc`, including the CCS auxiliary planes
@@ -61,6 +66,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
   Verified on a real Intel i915 CCS scanout.
 
 ### Fixed
+
 - `drmtap_grab_desc()` returns `-ENOTSUP` when the capture path yielded pixels
   but no transferable DMA-BUF (helper V2 fallback), instead of returning a
   descriptor the receiving process cannot convert.
@@ -68,12 +74,14 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
 ## [0.4.9] - 2026-07-20
 
 ### Added
+
 - Split-capture API for a privileged-export, unprivileged-convert boundary:
   `drmtap_open_render()` (render-node-only context, no KMS master or helper) and
   `drmtap_convert_dmabuf()` (detile and convert an exporter-supplied DMA-BUF to
   linear RGBA), plus the `drmtap_dmabuf_desc` descriptor.
 
 ### Changed
+
 - libEGL and libGLESv2 are loaded lazily with dlopen on first conversion instead
   of being linked, so a process that only grabs never maps the GPU userspace
   stack. They no longer appear in the `.so` DT_NEEDED or the pkg-config
@@ -82,6 +90,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
   (import once); steady-state conversion does no per-frame allocation.
 
 ### Fixed
+
 - Fast-grab cache-hit paths restage the plane layout, so a CCS import never runs
   with stale plane metadata.
 - `drmtap_convert_dmabuf()` validates the untrusted descriptor and rejects a
@@ -91,6 +100,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
 ## [0.4.8] - 2026-07-15
 
 ### Fixed
+
 - The per-thread EGL detile context leaked on `drmtap_close()`, causing OOM under
   a reconnect flap. It is now freed on close and at thread exit.
 - Wrong-CRTC selection: a `crtc_id == 0` / unbound connector was offered and
@@ -101,6 +111,7 @@ project share one version; the `libdrmtap` wrapper crate is versioned separately
 ## [0.4.7] - 2026-07-13
 
 ### Fixed
+
 - amdgpu EGL detile fix, privileged-helper hardening, and a batch of full-audit
   fixes.
