+++ docs/research/06_github_issues_analysis.md
@@ -244,4 +244,5 @@ From reading all these issues, the library MUST:
 - [x] Handle helper crash recovery — auto-respawn if helper dies mid-capture — done
 - [x] **Never CPU-deswizzle CCS modifiers** — `dumb_mmap` returns compressed data, only EGL can deswizzle. Return `-ENOTSUP` and require DMA-BUF fd + EGL (commit 7c7ce27)
 - [x] **V3 protocol: pass DMA-BUF fd via SCM_RIGHTS** for non-linear modifiers so parent can use EGL GPU deswizzle — shipped and now the default zero-copy capture path; Intel CCS detiles correctly in unprivileged/helper mode, with the V2 pixel copy as the fallback
+- [x] **Treat the convert descriptor as untrusted IPC input** — `drmtap_convert_dmabuf()` receives geometry and an fd from the other side of a process boundary. A descriptor claiming a frame larger than the fd actually backs made the CPU fallback mmap and read past the buffer, faulting the unprivileged converter with SIGBUS (a DoS). The fix requires a genuine DMA-BUF (a non-dma-buf fd is rejected; an immutable dma-buf cannot be truncated mid-read) and bounds the read against the buffer size from `lseek` (reliable across kernels, unlike `fstat` which reports 0 for a dma-buf before Linux 5.3), failing closed when the size is unknown. Covered by the `fuzz/fuzz_convert.c` libFuzzer target (0.4.12)
